+++ codec-http2/src/main/java/io/netty/handler/codec/http2/Http2ChannelDuplexHandler.java
@@ -81,6 +81,10 @@ public abstract class Http2ChannelDuplexHandler extends ChannelDuplexHandler {
         frameCodec.forEachActiveStream(streamVisitor);
     }
 
+    boolean hasActiveStream(){
+        return frameCodec.hasActiveStream();
+    }
+
     private static Http2FrameCodec requireHttp2FrameCodec(ChannelHandlerContext ctx) {
         ChannelHandlerContext frameCodecCtx = ctx.pipeline().context(Http2FrameCodec.class);
         if (frameCodecCtx == null) {
